refactor: lib/integration/ => lib/integrations/
