+++ src/components/modals/Agenda/AgendaModal.tsx
@@ -2,7 +2,7 @@ import { Modal } from "@mui/material";
 import React, { Dispatch, SetStateAction } from "react";
 
 import Agenda from "@/components/modals/Agenda/Agenda";
-import { classConfigRecurrentId, classRecurrentId } from "@/lib/integration/common";
+import { classConfigRecurrentId, classRecurrentId } from "@/lib/integrations/common";
 import { IntegrationConfig, RezervoClass } from "@/types/rezervo";
 
 const AgendaModal = ({
