+++ src/components/schedule/class/ClassPopularityMeter.tsx
@@ -3,7 +3,7 @@ import SpeedRoundedIcon from "@mui/icons-material/SpeedRounded";
 import { Tooltip } from "@mui/material";
 
 import RippleBadge from "@/components/utils/RippleBadge";
-import { isClassInThePast } from "@/lib/integration/common";
+import { isClassInThePast } from "@/lib/integrations/common";
 import { determineClassPopularity, stringifyClassPopularity } from "@/lib/popularity";
 import { ClassPopularity, RezervoClass, StatusColors } from "@/types/rezervo";
