+++ src/lib/integrations/sit.ts
@@ -1,7 +1,7 @@
 import { DateTime } from "luxon";
 
 import { GROUP_BOOKING_URL, TIME_ZONE } from "@/config/config";
-import { calculateMondayOffset } from "@/lib/integration/common";
+import { calculateMondayOffset } from "@/lib/integrations/common";
 import { SitDaySchedule, SitWeekSchedule } from "@/types/integration/sit";
 
 function sitScheduleUrl(token: string, fromISO: string | null = null) {
